requirements.txt
click
flask
line-bot-sdk
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
paho-mqtt
